feat: Optionally check Dafny CLI version before usage (#18)
This PR allows consumers of dafny.msbuild to optionally constrain the version of the Dafny CLI used for verification and compilation, using a version string (`DafnyVersion`), a Git commit hash (`DafnyVersionCommit`), or a Git tag name (`DafnyVersionTag`). If one is specified, then dafny.msbuild Tasks will check that the provided Dafny CLI matches the specified version. Using either a Git commit hash or tag name, requires that the `DafnyBinariesDir` property is set to a directory within a local dafny-lang Git repository.

Resolves #16.